Issue #61, PR #71: Corrected some inconsistencies.

+++ src/psyclone/tests/test_files/dynamo0p3/16.4.1_multiple_scalar_sums2.f90
@@ -39,7 +39,7 @@ program single_invoke
   ! Description: single kernel, multiple scalar sums & field writer argument
   use inf, only : r_def
   implicit none
-  integer(r_def)   :: rsum1, rsum2
+  real(r_def)   :: rsum1, rsum2
   type(field_type) :: f1
 
   call invoke( X_innerproduct_Y(f1, rsum1, rsum2) )
